Split run module
