+++ src/builtin/mod.rs
@@ -19,14 +19,22 @@ pub async fn run_fast_path(
     env_vars: Arc<HashMap<&'static str, String>>,
 ) -> anyhow::Result<(i32, Vec<u8>)> {
     match hook.repo() {
-        Repo::Meta { .. } => match hook.id.as_str() {
-            "check-hooks-apply" => meta_hooks::check_hooks_apply(hook, filenames, env_vars).await,
-            "check-useless-excludes" => {
-                meta_hooks::check_useless_excludes(hook, filenames, env_vars).await
-            }
-            "identity" => Ok(meta_hooks::identity(hook, filenames, env_vars)),
-            _ => unreachable!(),
-        },
+        Repo::Meta { .. } => run_meta_hook(hook, filenames, env_vars).await,
+        _ => unreachable!(),
+    }
+}
+
+async fn run_meta_hook(
+    hook: &Hook,
+    filenames: &[&String],
+    env_vars: Arc<HashMap<&'static str, String>>,
+) -> anyhow::Result<(i32, Vec<u8>)> {
+    match hook.id.as_str() {
+        "check-hooks-apply" => meta_hooks::check_hooks_apply(hook, filenames, env_vars).await,
+        "check-useless-excludes" => {
+            meta_hooks::check_useless_excludes(hook, filenames, env_vars).await
+        }
+        "identity" => Ok(meta_hooks::identity(hook, filenames, env_vars)),
         _ => unreachable!(),
     }
 }
