+++ src/cli/run/mod.rs
@@ -1,0 +1,7 @@
+pub use filter::{get_filenames, FileFilter, FileOptions};
+pub(crate) use run::{install_hooks, run};
+
+mod filter;
+mod keeper;
+#[allow(clippy::module_inception)]
+mod run;
